+++ doc/history.rst
@@ -26,6 +26,7 @@ History
   classification and regression problems.
 * Import and export graphs and their signals to NetworkX and graph-tool.
 * Save and load graphs and theirs signals to / from GraphML, GML, and GEXF.
+* Documentation: path graph linked to DCT, ring graph linked to DFT.
 * Merged all the extra requirements in a single dev requirement.
 
 Experimental filter API (to be tested and validated):